[FEAT] Add PlayerExperienceSystem And BinderUI

+++ Assets/01_Scripts/ECS/Authoring/Enemy/EnemyAuthoring.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public class EnemyAuthoring : MonoBehaviour
 {
     public float Speed = 5f;
-    public float ExperienceAmount = 1;
+    public int ExperienceAmount = 1;
 }
